+++ onlyofficeeditorapi.php
@@ -18,7 +18,7 @@
  * The ONLYOFFICE editor api.
  *
  * @package    mod_onlyofficeeditor
- * @copyright  2023 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright  2024 Ascensio System SIA <integration@onlyoffice.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */